+++ collab-database/tests/database_test/view_test.rs
@@ -224,7 +224,7 @@ async fn duplicate_database_view_test() {
 #[tokio::test]
 async fn duplicate_database_data_serde_test() {
   let database_test = create_database_with_default_data(1, "1").await;
-  let duplicated_database = database_test.get_all_database_data();
+  let duplicated_database = database_test.get_database_data();
 
   let json = duplicated_database.to_json().unwrap();
   let duplicated_database2 = DatabaseData::from_json(&json).unwrap();
