+++ collab-database/tests/user_test/mod.rs
@@ -1,3 +1,4 @@
+mod async_test;
 mod cell_test;
 mod database_test;
 pub mod helper;
